quiet color scheme
